extract main feedback logic into a hook and share it between the floating, and button based feedbacks
