+++ static/app/components/feedback/widget/feedbackWidgetButton.tsx
@@ -1,36 +1,13 @@
-import {useEffect, useRef} from 'react';
-import {Feedback, getCurrentHub} from '@sentry/react';
+import {useRef} from 'react';
 
 import {Button} from 'sentry/components/button';
+import useFeedbackWidget from 'sentry/components/feedback/widget/useFeedbackWidget';
 import {IconMegaphone} from 'sentry/icons/iconMegaphone';
 import {t} from 'sentry/locale';
-import ConfigStore from 'sentry/stores/configStore';
-import {useLegacyStore} from 'sentry/stores/useLegacyStore';
 
 export default function FeedbackWidgetButton() {
   const buttonRef = useRef<HTMLButtonElement>(null);
-  const config = useLegacyStore(ConfigStore);
-  const hub = getCurrentHub();
-  const feedback = hub.getIntegration(Feedback);
-
-  useEffect(() => {
-    if (!buttonRef.current || !feedback) {
-      return undefined;
-    }
-
-    const widget = feedback.attachTo(buttonRef.current, {
-      colorScheme: config.theme === 'dark' ? 'dark' : 'light',
-      formTitle: t('Give Feedback'),
-      submitButtonLabel: t('Send Feedback'),
-      messagePlaceholder: t('What did you expect?'),
-    });
-
-    return () => {
-      if (widget && feedback) {
-        feedback.removeWidget(widget);
-      }
-    };
-  }, [config.theme, feedback]);
+  const feedback = useFeedbackWidget({buttonRef});
 
   // Do not show button if Feedback integration is not enabled
   if (!feedback) {
